api/requirements.txt
fastapi
uvicorn
psycopg2-binary  # Driver para conectar Postgres
websockets
passlib==1.7.4
bcrypt==3.2.2
PyJWT # Para generar y descifrar los tokens JWT.
python-multipart #FastAPI lo necesita para leer los formularios de login estándar.
python-dotenv